% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SomethingCool.R
\docType{package}
\name{SomethingCool}
\alias{SomethingCool}
\alias{SomethingCool-package}
\title{Infrastructure Tools Package for SomethingCool}
\description{
This package collects various scripts and templates
we've found to be useful in organizing projects.
}
